// interfaces/Posts.ts
// Generated by https://quicktype.io

export interface PostResponse {
    data: DatumPosts[];
    meta: Meta;
}

export interface DatumPosts {
    id:         number;
    attributes: DatumAttributes;
}

export interface DatumAttributes {
    title:       string;
    description: string;
    slug:        string;
    createdAt:   string;
    updatedAt:   string;
    publishedAt: string;
    image:       Image;
}

export interface Image {
    data: Data;
}

export interface Data {
    id:         number;
    attributes: DataAttributes;
}

export interface DataAttributes {
    name:              string;
    alternativeText:   null;
    caption:           null;
    width:             number;
    height:            number;
    formats:           Formats;
    hash:              string;
    ext:               EXT;
    mime:              MIME;
    size:              number;
    url:               string;
    previewUrl:        null;
    provider:          string;
    provider_metadata: ProviderMetadata;
    createdAt:         string;
    updatedAt:         string;
}

export enum EXT {
    Jpg = ".jpg",
}

export interface Formats {
    small:     Medium;
    medium:    Medium;
    thumbnail: Medium;
}

export interface Medium {
    ext:               EXT;
    url:               string;
    hash:              string;
    mime:              MIME;
    name:              string;
    path:              null;
    size:              number;
    width:             number;
    height:            number;
    provider_metadata: ProviderMetadata;
}

export enum MIME {
    ImageJPEG = "image/jpeg",
}

export interface ProviderMetadata {
    public_id:     string;
    resource_type: ResourceType;
}

export enum ResourceType {
    Image = "image",
}

export interface Meta {
    pagination: Pagination;
}

export interface Pagination {
    page:      number;
    pageSize:  number;
    pageCount: number;
    total:     number;
}